add authentification

+++ routers/posts.py
@@ -6,8 +6,9 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.orm import selectinload
 
 import models
+from auth import CurrentUser
 from database import get_db
-from schemas import PostResponse, PostCreate, PostUpdate
+from schemas import PostCreate, PostResponse, PostUpdate
 
 router = APIRouter()
 
@@ -18,18 +19,16 @@ router = APIRouter()
     response_model=PostResponse,
     status_code=status.HTTP_201_CREATED,
 )
-async def create_post(post: PostCreate, db: Annotated[AsyncSession, Depends(get_db)]):
-    result = await db.execute(select(models.User).where(models.User.id == post.user_id))
-    user = result.scalars().first()
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="User not found",
-        )
+async def create_post(
+    post: PostCreate,
+    current_user: CurrentUser,
+    db: Annotated[AsyncSession, Depends(get_db)],
+):
+
     new_post = models.Post(
         title=post.title,
         content=post.content,
-        user_id=post.user_id,
+        user_id=current_user.id,
     )
     db.add(new_post)
     await db.commit()
@@ -67,6 +66,7 @@ async def get_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]):
 @router.put("/{post_id}", response_model=PostUpdate)
 async def update_post_full(
     post_id: int,
+    current_user: CurrentUser,
     post_data: PostCreate,
     db: Annotated[AsyncSession, Depends(get_db)],
 ):
@@ -81,19 +81,14 @@ async def update_post_full(
             status_code=status.HTTP_404_NOT_FOUND, detail="Post not found"
         )
 
-    if post_data.user_id != post.user_id:
-        result = await db.execute(
-            select(models.User).where(models.User.id == post.user_id)
+    if post.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not autorized to update this post.",
         )
-        user = result.scalars().first()
-        if not user:
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND, detials="User not found"
-            )
 
     post.title = post_data.title
     post.content = post_data.content
-    post.user_id = post_data.user_id
 
     await db.commit()
     await db.refresh(post, attribute_names=["author"])
@@ -104,6 +99,7 @@ async def update_post_full(
 @router.patch("/{post_id}", response_model=PostUpdate)
 async def update_post_partial(
     post_id: int,
+    current_user: CurrentUser,
     post_data: PostUpdate,
     db: Annotated[AsyncSession, Depends(get_db)],
 ):
@@ -118,6 +114,12 @@ async def update_post_partial(
             status_code=status.HTTP_404_NOT_FOUND, detail="Post not found"
         )
 
+    if post.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not autorized to partially update this post.",
+        )
+
     update_data = post_data.model_dump(exclude_unset=True)
     for field, value in update_data.items():
         setattr(post, field, value)
@@ -129,7 +131,11 @@ async def update_post_partial(
 
 # DELETE POST #
 @router.delete("/{post_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]):
+async def delete_post(
+    post_id: int,
+    current_user: CurrentUser,
+    db: Annotated[AsyncSession, Depends(get_db)],
+):
     result = await db.execute(
         select(models.Post)
         .options(selectinload(models.Post.author))
@@ -139,6 +145,12 @@ async def delete_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]
     if not post:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND, detail="Post not found"
+        )
+
+    if post.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not autorized to delete this post.",
         )
 
     await db.delete(post)
